misc/HtmlTestSuite/results/test-0959.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head>
        <title> scheduler: stylesheets blocking scripts document.write</title>
        <script src="/resources/testharness.js"/>
        <script src="/resources/testharnessreport.js"/>
        <script src="testlib/testlib.js"/>
</head>
<body>
        <div id="log">FAILED (This TC requires JavaScript enabled)</div>
        <div id="test">Test</div>

        <!-- this stylesheet blocks scripts -->
        <script>
          test(function() {
            document.write("&lt;link rel='stylesheet' href='css/background.css?pipe=trickle(d2)'&gt;");
            // note that the pass condition here is not per spec (but does match implementations) as of 2012-06-26
            assert_equals(getComputedStyle(document.getElementById("test")).position,
                          "static");
          });
        </script>
</body></html>
